requirements.txt
streamlit
openai
langchain
python-dotenv
docker
transformers
sentence-transformers
langchain-community
openpyxl
pandas
googlesearch-python
